requirements.txt
youtube_transcript_api
streamlit
google-generativeai
google-cloud
google-cloud-speech
python-dotenv
pathlib

-e.